dns-policy-app/app/requirements.txt
Django==5.1.6
psycopg[binary]==3.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
